etl/requirements.txt
pymongo
sqlalchemy
pandas
vaderSentiment
psycopg2-binary
